Removes completely core.py by refactoring methods into ventilation.py
