+++ archetypal/__init__.py
@@ -27,6 +27,5 @@ from .dataportal import *
 from .plot import *
 from .trnsys import *
 from .template import *
-from .core import *
 from .umi_template import *
 from .cli import *
